requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
executing==2.0.1
faiss-cpu==1.8.0
fastjsonschema==2.19.1
filelock==3.15.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.2.3
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.77
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nodeenv==1.9.1
numpy==1.26.4
openai==1.34.0
orjson==3.10.4
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
tenacity==8.3.0
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.26.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.9.4
